requirements.txt
# AI/ML Framework
google-genai==1.50.1

# Environment & Configuration
python-dotenv==1.0.1

# Web Scraping & Search
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
google-search-results==2.4.2

# Web Interface
streamlit==1.51.0

# PDF Generation
reportlab==4.4.5

# Data Visualization
matplotlib==3.10.7
seaborn==0.13.2

# Data Processing (required dependencies)
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0